[chore] layout 함수 Extension 파일로 분리
